pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["s.burovin <s.burovin@tinkoff.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.3.1"
torch = "^2.0.1"
pyyaml = "^6.0.1"
hydra-core = "^1.3.2"
matplotlib = "^3.8.0"
tqdm = "^4.66.1"
pandas = "^1.3.3"
pre-commit = "^3.4.0"
dvc = {extras = ["gdrive"], version = "^3.30.3"}
mlflow = "^2.8.1"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
tritonclient = {extras = ["http"], version = "^2.41.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
